+++ android/app/src/main/java/es/movingout/app/MainActivity.java
@@ -1,0 +1,5 @@
+package es.movingout.app;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}